adjusted Constructor for GraphHandler implementation
removed shortestPathList attribute, since only a single call is
necessary. saving as attribute is redundant.

also added a recursive check for checkFileExists method.
it now checks again in case the file already exists and rename is
chosen. before it was possible to enter a new name which also already
exists without another check.