adjusted GraphWriter
now implements Runnable interface

added a boolean createFile
createFile is checked by run(). if set to false, no file is created.

set checkFileExists to public.
the method now is called by main, which allows multithreading of file
creation and console output.
The method now also is responsible for a possible set of createFile to
false.